Make sure all tile data is available if not in search results
